Handle non integers
I guess it's possible to get a half mark

+++ js/tables.js
@@ -61,7 +61,7 @@ function compareCells(a, b) {
     var a = $.text([a]);
 
     if (isNumber(a) && isNumber(b)) {
-        return parseInt(b) - parseInt(a);
+        return parseFloat(b) - parseFloat(a);
     } else {
         return a.localeCompare(b);
     }
